refactor(ui): slide component
