+++ ui/slider/src/slide/slide-image/index.ts
